<Altered getAlertsForStation>
Modified the sql queries in getAlertsForStation function to now return
all active alerts for all stations, but then update all active alerts
for the station whose ID was passed to the function so that those
alerts are no in-active. As this function is called when someone swipes
their card at a station, and the station passes its ID to the servlet,
it is assumed that once someone swipes in at a station that they will
deal with any active alerts at that station, thus they can marked in
the system as taken care of.

#Changed sql statements in getAlertsForStation.