<Added Sign Recording and Retrieving Functionality>
Added code to getAlerts Functions that adds the tagID and stationID
passed from the app to the sql database along with a timestamp,
creating a record of when what tag were scanned at what stations.

Added getLatestSignins() function to return the latest signing for each
tagID from the database.

+Added to code to getAlertsForStation to store the tagID and stationID
to the signins table with a timestamp.
+Added to code to getSecurityAlertsForStation to store the tagID and
stationID to the signins table with a timestamp.
+Added getLatestSignins() to return the details of the latest signin
for each tagID.